requirements.txt
streamlit
openai
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
langchain
httpx
